Fixed comma operator precedence
Comma operator used to have a higher precedence than assignment
expression, this disallowed expressions like: `x=z+1, y=z+2`. To fix this comma operator is given the lowest precedence.

This though introduces surprising behavior to the user.
In a variable declaration statement like `var x = 1, 2, 3;`, the variable initializer is a comma operator expression thus evaluating to 3, `var x = (1, 2, 3)`.
In an assignment expression, the assignment has a higher precedence than the commamaking this expression `x = 1, 2, 3` evaluate to `(x=1), 2, 3`.
This should be clear when treating variable declaration as statements as they should be.