Function declaration and execution support
Defined new AST types for function declaration, and new Function class to
represent the language's functions at runtime, lastly added Parser,
Interpreter and AstPrinter support

Currently all function calls are executed as if the function is always
globally defined.
In other words, any parameters and variables defined in the function are defined in an environment that's a direct child of the global environment.