Enabled declaring a variable more than once in global scope
Prohibiting declaring a variable more than once introduced some errors
in functionality where Resolver and interpreter got out of sync.
```
var a = funThatTakesNoArgs(SomeArgument);
var a = fun();
[line 1] Error at 'a':
Variable a already declared in this scope
a = method();
Assignment to undefined variable 'a'
[line 1]
```
Easiest way to handle this was to allow multiple declaration of the same
variable in global scope. This also made the repl session more pleasant to work with as multiple declaration of the same variable (specially functions and classes) are desired in repl sessions.